Create video functionality for Android
Substantial refactor of video functionality to accommodate limitations
on Android. In essence, instead of manipulating time for a singular
VideoPlayer component, for Android we need to have two instances of a
VideoPlayer, one to play as normal and the other to play in reverse. To
do this, I abstracted the functionality to step through a video out of
PlayableVideoPlayer and into a PlayableVideoPlayerController, which contains
a reference to both the videos and is targeted by the LerpVideoPlayerTime
playable accordingly.

In the meantime, refactored MeshRendererUpdater and various other methods
to help with debugging, while also reorganizing the code and revising the
relevant prefabs. We also now have a spinner loading icon and related events.

Next steps are to do a code cleanup, namely removing all video buffering
functionality (this does not work), add more Android-specific code for the
LerpVideoPlayerTime playable, and do other miscellaneous cleanup.