Refactor localization; create new Modify directory with related functionality
The new Modify directory contains functionality to create
variation within a scene. For example, localization falls
under this category - the scene largely stays the same,
but the localization tools here allow us to change the text
as needed. I've also refactored the PopulateText tool to
accommodate the new logic; all LocalizedTexts take reference
to a LocalizationCorpus, a set of texts that should relate to a
particular scene  (i.e. Chapter 1, Chapter 2), and each corpus
contains all the translations for that scene.

The translations are grouped under a new Language scriptable object
type, which are used as an easily referenceable item in the hierarchy,
but can also contain attributes of their own (for example, they can
be tied to a System.Language, which can allow us to specify defaults
and such as the system develops).

Another important point to note is the Layout scriptable object type.
I envision these Layouts being related to Languages; for example,
a scene in Mandarin might require a slightly different setup than a scene in
English. I haven't fully developed this idea yet, but the core
idea is there to be built upon whenever needed.

ModifyHandler is the scene object that will allow us to update any
localized texts when in game. It communicates with the LocalizedTexts
using a ComplexEvent called UpdateLanguage.

LocalizationManager is now obselete - will delete it next. Next step: Use the
localization to switch between cyberspeak and plain English in Violino.