Refactor all ResponsiveElements; create SerializableElementReference; refactor Modify functionality and create new Modify tools
Finished a huge refactor in which all ResponsiveElements now
inherit from SerializableElement. This means that all ResponsiveElements
are now, be default, hooked into the Modify functionality, allowing
us to save any data from a ResponsiveElement based on a current Layout.

Importantly, every SerializableElement creates a custom ID for itself,
which we use when creating a file for its data, as well as to allow for flexibility
in referencing the elements from saved JSON if need be (there was
no easy way to do this otherwise, as game objects in Unity do not
persist their instance IDs in between sessions).

Included in these changes is a new SerializableElementReference script,
which allows us to store references to game objects in JSON in conjunction
with the ReferableElement class. This is used by the new
ResponsiveToggleController, which allows us to toggle game objects on and off.

All ResponsiveElements now, be default, should incorporate support
for breakpoints. Ideally when creating any new responsive functionality,
we should create it once, the right way, with support for breakpoints,
so that any projects we build using that functionality don't lose
layout because of serializer mixups.

That being said, I created a new LayoutTools editor window and revised
the TextTools editor window in order to improve the workflow. With the
new LayoutTools in tow, it means that the ModifyHandler can exist as a
persistent singleton created, along with App Navigator and Detect Screen
Size, in the bootstrap scene.

Additionally, Modify functionality, including language, layout, and the
Modify Handler, were refactored to be more flexible and generic. For
example, different language options are now specified via the TextFamily
script, owing to the fact that Modify can be used not just for
localization, but to execute arbitrary text and layout changes
for storytelling purposes as well.

Lastly, I revised the ComplexEvent so that, when raised without any data,
it will trigger all of the listeners using a blank event payload. When
combined with null reference checks on the listener end, this allows
for maximum flexbility in creating complex behaviours.

That should cover most of the big changes. Next step is to create a handler
to save and read arbitrary data in UserPreferences. This should hook in
nicely with the UnityEventTrigger script (also new), which allows us to
execute arbitrary Unity events based on BoolVariables (am planning to extend
this to allow for more types of conditions as well).