Add logic for Tutorial snap elements
Added new ScrollSnapController, which allows us to create tutorial-like
panels to give instructions to the user. It's very much not as modular
and generic as it could be, but it's a good start.

It's meant to work in conjunction with a very specific game object
setup, so next step is to create a prefab out of it.

Other than that, modified some standard AltSalt scriptable objects
in order to make the tutorial sync correctly with the rest of the
project. Also, created a new DirectorTools script to allow for
minute handling of a PlayableDirector if needed.