Add ConditionResponse logic; add UserDataController; reorganize several folders
Added a new system to allow for complex interactions between
systems without using any code. All of this code has been placed into
_AltSalt/__Core/Data Types/Events/ConditionResponse/.

The idea here is that, based on certain conditions of the game state,
we want to be able to execute responses, or events, to modify dependencies
like game objects and scriptable objects accordingly. I created a
base ConditionResponse script, and then created several derived types
to provide handling for Bools, Floats, Text Families, and Layouts.

In most cases, to make use of these ConditionResponses, we will need to use the
related classes ConditionResponseTrigger and ConditionResponseTriggerBehaviour.
The idea here is that the ConditionResponseTrigger is a way of keeping all
of our possible functionality handy in one place, which makes for a good
user experience when building stories.

Since we need the ConditionResponseTrigger in not just MonoBehaviours, but
also inside of our Playables, I created ConditionResponseTriggerBehaviour
separately as a way of maximizing flexibility into the future.

ConditionResponseTriggerBehaviours can be executed on start, or they can be triggered
by other ConditionResponseTriggers, or by UI, or basically any item in the
scene that can referenece game objects. Used in conjunction with
ComplexEventPackagerBehaviours, this allows us to create robust, flexible, and
(I hope) scalable and generic architecture that can be used in a variety of
situations.

For example, you can sequence a series of ConditionResponseTriggerBehaviours
to enable or disable components and game objects in the scene dynamically based
on arbitrary BoolVariables or FloatVariables.

Lastly, the UserDataController now allows us to take BoolVariables, StringVariables, or
FloatVariables and save that data to PlayerPrefs. All we need to do is raise the
complex event, StoreData, with a scriptable object that we'd like to store. The
UserDataController should also contain a reference to that scriptable object - this so
we can both keep track of what data we'd like to persist, and so we can save that data
effectively - and that data will be saved. Data can also go in the other direction and be
read from PlayerPrefs to set the game's BoolVariables, FloatVariables, etc.
(I haven't made the reading part very robust yet, for now I'm just
doing it once when the game loads, but can make that more flexible to work
on the raising of a simple event soon).

Next thing to do is writing a Playable for our new ConditionResponse functionality,
as well as creating the robustness for the reading of data from UserDataController.