Refactor portions of Serializable Element and Layout functionality; add beginning of audio and draggable functionality
It turns out that the saving of layout data to JSON files was not
working between sessions because IDs were not persisting as intended,
also when more than one scene was active, it would save the files
into the incorrect location. So I revised the functionality to
handle those scenarios. The TextTools and LayoutTools editor windows
were also not triggering events properly during use, so I fixed that.

I got started on the beginnings of our audio functionality, including
the ability for any item to trigger the creation of an AudioSource
using a ComplexEvent, passing a new AudioClipBundle as a parameter.
An AudioClipBundle is a scriptable object that contains reference
to one or more audio clips. In this way, we can allow elements within
a scene to, in particular, trigger audio clips that will persist
during scene transitions.

I started working on drag functionality under the impression that I
would need to move the tutorial completely out of the UI in order for
it to work. Turns out that the UI implementation still runs at 60fps
on an iPhone 6s, and that it's still functionaly on a Galaxy s5, which
means that it's good enough. For some reason FPS drops to 50 on an iPad,
so may still refactor this in the future, but it's not a priority right now.
In any case, the draggable functionality is in beginning stages and can
be built upon more if and when we start deciding to explore more
kinds of touch interactions.

Next step is to clean up the parent Assets/ git module, and then begin
a refactor / overhaul of the event system to make it more robust and
scalable.