Work in progress - Refactor event system to allow logging of callers
Am currently making an intermediate layer between SimpleEvents/ComplexEvents
and how they are called. The layer consists of SimpleEventTriggers and
ComplexEventTriggers, which force us to pass in the calling object
whenever we activate the trigger - the event in question thus has
a reference to the calling object, which we can log into the console.

As a result, we now have a traceable record of all events fired, from where,
and the listeners that are activated as a result. This is important
because as scenes become more complex, we'll need to deal with more
and more interactions (it's almost unwieldy with a small project the
size of violino), so we want to be able to output the logs and get
a sense of exactly what's happening in the app without too much effort.

Also started working on some unit tests - will be aiming to implement
more of these.

Also worth noting is that SimpleEvents and ComplexEvents can still be
raised without a trigger, but the StoreCaller() method must be used
to pass in the referencing object before the event can be raised. That
means there's a little flexibility here in how events get triggered,
but the important thing is that we log all the interactions that
are taking place whenever we want.

Next steps are to iron out any kinks in the system, in particular
adding the active scene name to the logs for even better tracing,
followed by perhaps adding a similar system for all variable
references (BoolReference, FloatReference, etc.), and the creation
of a set of logs that will keep track of scriptable object
dependencies overall. The other day, I opened an old project and ran into a
bug wherein all of the serialized data had disappeared - and I had
no way of knowing how to piece together the web of interactions and
dependencies that the project required.

With the logs, I'm hoping we can have a backup of all those references
in order to, if not automatically repopulate references, have a map to
allow it to happen relatively easily. When it comes to the need to rebuild
and maintain projects, it's not a matter of if, just when, that we'll lose
all of the serialized data either due to bugs, Unity updates,
or refactoring on AltSalt's part.