(Work in progress) - Event system refactor to log callers and beginning of RegisterDependency functionality
Substantial changes to the event system, namely with the purpose
of being able to use ComplexEventTriggers and SimpleEventTriggers
as intermediates between raising complex events and simple events
from game objects. This is because having the intermediaries
allows us to insert additional logic before the event is raised
- namely, complex and simple events can now only be raised
when the calling game object is first registered to the event
so that it, as well as its scene, can be logged to the
console if we ever want to debug.

This is important because, as scenes grow more complex, it'll be important
to be able to find how certain events are firing, in particular from
which objects. In this way, we should be able to continue building
confidently with the knowledge that if a project ever changes hands or
requires maintenance, that we will be able to activate the logs and
get a good sense of all the communication that is happening in a scene
between game objects and events.

The ConditionResponse functionality was treated similarly - you can
activate a boolean in AppSettings that will output all of the triggers,
and their conditions, when they are raised, so we can get an even
deeper understanding of the project's logic if need be.

Note: This means that UnityEvents should NEVER be used to execute
functionality on game objects without first passing through a SimpleEvent,
ComplexEvent, or ConditionResponse. Otherwise, the functionality is
untraceable and could get lost.

The last piece of this puzzle includes keeping track of all dependencies
in text files so that, if we ever lose our serialized data, we can
relatively easily reconstruct a scene's functionality by looking at the
logs. This functionality is currently in progress and may be broken -
I'm planning on creating a registry of all dependencies inside of
/__Project/z_Dependencies. The idea is that we will have logs of every
scriptable object event and variable, which will show the game objects
that are dependent on them by scene name. We will also have logs of
every game object in the project, organized by scene, with a record
of all of their scriptable object dependencies.

Next step is to refactor all scriptable object variables to inherit
from the same abstract VariableBase class - in this way, I'm hoping to make them
easier to work with as I continue the refactor.