Work in progress - Create new editor tool for populating dependencies
Seem to have gotten the dependency recorder working. Using recursion
and careful type checking in _AltSalt/Core/Editor/RegisterDependencies,
we can now drill down to any scriptable object references that we're
most concerned with that will cause most or all of our scaling issues
if we ever need to maintain a project. It saves these references to
__Project/z_Dependencies.

As a bonus, I've also managed to record the targets
and methods on any UnityEvents we might have within components.

Worth noting: In order to make the RegisterDependencies functionality
more or less maintainable, I needed to introduce a few empty base classes
like TriggerBase, VariableBase, and VariableReferenceBase
in order to easily reference their types via reflection (EventBase and
ConditionResponse were  already in use). The only issue
is that any new classes we create that make use of scriptable object
references should also be added to the RegisterDependencies functionality,
or inherit from one of the base classes already specified.

I can foresee that this script might be a headache to maintain, but maybe
not. It works well for now and will hopefully be fairly robust.

Next step is to do code cleanup, add suport for PlayableAssets, and
make sure it works across multiple scenes.