Finish DependencyRegistration functionality; add a lot of new base classes
Finished a pass at the RegisterDependencies functionality. The idea here is
that you can use the new tool under Tools/AltSalt/RegisterDependencies to
search the codebase for all scriptable objects and register the components
on which those objects are used. In passing, it also records data from
UnityEvents and conditions from ConditionResponse triggers.

By doing this, we can now create records of functionality to help with
maintenance and debugging of projects. By having an overview and record
of all event and scriptable object logic, the idea is that we mitigate the
largest flaw of the architecture we have in place, namely that before it was
hard to keep track of dependencies, making it hard to scale projects.

To do this, I introduced a lot of new base classes to act as glue so that I
could parse the fields. These new base classes may be useful in the future as
well.

Next steps are to create integration tests for the RegisterDependencies functionality,
as I feel this has the potential to be brittle during large code refactors.

Also, I'm going to reorganize and rename folders to more accurately reflect the
new developments in architecture over the past several commits.