Add single event argument option for SimpleEventTrigger
In order to allow for a single event trigger to activate more than one
event, I added an option that allows you to specify a target event.
If the event exists in the trigger's list, it will activate. This is
handy in the case of the FadeElement, for example, which must trigger
both the FadeInCompleted event and the FadeOutCompleted event, so this
allows us to have just one trigger handle both those events.