WORK IN PROGRESS - Begin work on PageBuilder and TimelineUtilities tools
Substantial work on creating tools to speed up story creation process.
Everything should be largely self-explanatory. Everything is being
created inside of the EditorTools folder, and I'm breaking up the scripts
as needed in order to maintain readability. Not sure what the best user
experience is, so I'm revising as I go.

Last commit before a refactor that could potentially break large parts
of the project.

Note that the pivot point of a background sprite was changed, so that may
break the positioning of some sprites in Violino. Also, there are some
miscellaneous assets and perhaps scripts in this commit that need to be
refactored out.