WORK IN PROGRESS - Revise Timeline Utitilies; prepare for refactor
Seem to have gotten the sequencing functionality working. In other
words, it's now possible to easily set clips to occur sequentially;
to do this, I just organize the selected clips based on the order
of their parent tracks. This can also be done in reverse.

Also refactored the Select Next and Select Previous functionality.

Next step is to make it so that it's possible to transpose all clips
following your selection after making any edits. It's currently done
haphazardly in some functions but I'd like to clean up the code to
make this work better and more uniformly. This could potentially
break things so I'm committing now just in case.