Implement TimelineClipTools using new UI Elements functionality
Finished migrating what I have so far of the TimelineClipTools over
to the new functionality, using UXML and USS files, the whole deal.

Next step is to implement the Expand and Expand (<>) functionality,
as well as fix a couple of small issues with the other methods. Then,
move onto the rest of the creation and page builder tools.