Finish moving PageBuilder editor functionality to UI Elements
I've been working with UI Elements for a few days now and seem
to have gotten the functionality and display logic redone using
the new functionality. PageBuilderWindow and TimelineUtilitiesWindow
contain all of that presentation logic, and I'm trying to move all
of the functional logic outside to other files (not there yet).

Next step is a big folder restructuring to see if I can reorganize
everything so that it's easier to maintain / makes sense.