Fix visual styling of editor tools; add functionality for adding breakpoints
Did some housekeeping and introduced CSS classes to allow for detailed
styling of UXML elements, such as removing margins, and implemented in a few
templates. Also started with the editor tools responsive functionality. Plan
is to refactor the ResponsiveElement class so that we can execute responsive
actions based on priority, that way we can make sure that elements can execute
their responsiveness in specific order if need be.

Last commit before the refactor.