Add ResponsiveUtils window; clean up responsive functionality
In order to speed up the process / workflow of adding responsive functionality,
created a new editor window, ResponsiveUtils, to add aspect ratios, set breakpoints,
and trigger screen resizes if need be. Also took the opportunity to fix a lot of bugs
related to responsive functionality, clean up the code, and add the ability to crop
the camera view using ResponsiveAutoViewport.

Basically, DetectScreenSize is now responsible for setting variables related specifically
to the *device* width, height, and aspect ratio, and PrepareScene uses those values to set
the width, height, and aspect ratio of the *scene* based on user preferences.

This is really handy, because it allows different scenes or stories within an app to
have different aspect ratios, and also allows us to force letterboxing, pillarboxing, or
cropping for certain aspect ratios, which means we can ensure the projects look great on
all devices while also cutting down on the uncertainty and work of developing for every
screen size in existence.

Other than that, fixed the inheritance scheme and enabling of ResponsiveElements so that it's
cleaner and makes sense. Importantly, PrepareScene is now also responsible for executing
responsive actions, not the ResponsiveElements themselves. This is because I realized it's
important to be able to set the priority of responsive actions in case one is dependent on the
other; not only that, but it's cleaner and more maintainable to have a separate entity
handling all of that. (To do this, ResponsiveElements register themselves to PrepareScene
using the new complex event, ResponsiveElementEnable).

Also added properties to Simple and Complex events that will allow us to see their
dependencies easily in the inspector. Will continue populating those as development continues.
Next steps are to add in housekeeping logic for ResponsiveElements (such as registering Undo
actions and  adding Saveable interfaces), as well as finally start working to finish
Riveted.