WORK IN PROGRESS - Miscellaneous tools fixes; begin editor tools for responsive timeline clips
Improved on a lot of functionality, including fixing a bug wherein
editor tools were erroneously changing values due to their editor windows
being bound to data incorrectly, as well as added lots of utilities inside
of SimpleEventTrigger, ComplexEventTrigger, and SimpleEventListener to
allow for their usage in script. The utility additions were due to
encountering hiccups in workflow while populating texts, namely being
unable to trigger layout and text changes from the TextTools and LayoutTools.

A big change that's currently underway is implementing editor tools for responsive timeline clips.
To do this, I refactored ResponsiveElement to now use the new improved IResponsive interface,
which is also implemented by ResponsiveLerpToTargetBehaviour. Unity's Timeline
implementation isn't as clean as it could be unfortunately, so there's some
code bloat in the ResponsiveUtils window to deal with different kinds of
responsive clip class types. I also needed to add references to the playables'
clip and track assets via ResponsiveLerpToTargetTrack. The scripts related to
clip responsiveness are now pretty dependent on one another unfortunately and
verbose. Everything works for now and is relatively clear to understand; however,
should keep an eye on this functionality and potentially see if it can be
refactored in the future.

Currently the ResponsiveLerpToTargetBehaviour timeline script uses a SimpleEventListener
to respond directly to changes in screen size. Before wrapping this set of functionality up,
I'm planning to refactor PrepareScene so it can handle that functionality instead,
that way we don't have two different implementations for responsiveness in
the codebase.

Next steps are to do the PrepareScene refactor, then finish up the
EditRectTransformResponsivePosClip editor tool.