Finish editor tools for responsive timeline clips; fix playable bugs
Added lots of connection logic to allow the page builder tools to connect
with the responsive playables. The responsive playables now hook up to
PrepareScene, which allows us to execute responsive actions in the editor
whenever we resize the screen. The editor tools also allow us to add breakpoints
as needed while working with the responsive clips and seem to work as intended.

Importantly, I made it so that the  ResponsiveRectTransformPos and ResponsiveRectTransformScale
playables now make use of same ResponsiveVector3Behaviour and ResponsiveVector3Clip.
The code between them is identical; the same approach can almost definitely be taken
with the rest of the playables; however, this would break serialization on our
current projects. That being said, would like to revisit this once I hook the playables
up to the Modify functionality (namely, layouts).

Other than that, fixed code on various playable tracks that were erroneously flagging
attributes on components that they didn't need to, which was causing bugs with the
workflow when trying to check different screen sizes (elements were rendering with
the incorrect breakpoint information, etc.)

Also added lots of miscellaneous logic to the ComplexEventTrigger, ComplexEvent, and
EventPayload to allow for the responsive elements to send themselves as parameters to
PrepareScene. We now have the ability to send any System.Object as a parameter via complex
events.

Next steps are to continue with miscellaenous cleanup as I work towards finishing
Riveted in the Word.