WORK IN PROGRESS - Add robust handling for handling dynamic audio
Added a new playable, AudioLerpSnapshot, in order to transition
audio mixers between snapshots. Importantly, I've added handling for
crossfading as well as standard transitions. It's worth mentioning
that snapshots typically transition on a logarithmic scale, so the
behaviour when attempting to do a simple crossfade between snapshots
typically won't work as expected and audio will fade in / out too quickly. For
that reason, the AudioLerpSnapshotClip contains an option for
'crossfade' which uses a separate equation to make sure behaviour
in clips performs as expected.

Other than that, made the AudioFadePlay playable more robust to
handle multiple audio sources, so we can easily sync clips up whenever
we want to have a dynamic soundtrack. This makes use of exposed references,
which hold references to audio source game objects in the scene.

Also important: I've decided that the workflow for audio should be to have
a separate mixer for each audio track. This allows us to have granular control
over each element in a mix, crossfading and adjusting levels as needed on the fly.
In addition, each of these mixers should hook up to a master mixer, where we can
also have flexibility in manipulating groups of mixers and snapshots.
As far as I know, whenever a mixer drops to -80 decibels, the system
also stops processing the audio signal, which will be important
for optimizing performance on mobile (other strategies would have resulted
in the audio signal continuing to be processed).

AudioClipBundle has been revised as a result. Any dynamically created audio
sources should contain a reference to the mixer group it should output to.
In this way, we can manipulate the audio mixers with fine levels of details,
across scenes, without requiring a reference to the audio source itself.