Add Touchable; modify Audio playables; add TextUtils; various other fixes
Biggest / most important change is the addition of a Touchable script
to make things interactive that works with the resizable collider script.
In addition, fixed the bounding box utility behaviours that allow us to
visualize rect transforms and text boxes in the editor.

While implementing audio and dealing with a few real-world use cases
with Violino, also made various small fixes and added utility functions
to allow for more flexibility.