Add selection tools to CreateCoreElements; miscellaneous fixes
Biggest change is that I added util functions to select children
and save selections to the EditorTools. This is important, as it
allows us to work with / manipulate large hierarchies.

Other than that, made several miscellaeneous fixes, including...
-Removing execution of responsive actions from PrepareScene in Enable(), which
was causing game objects to break (it was too early for them to initialize and
this is handled by the Start() method anyway)
-Removing a conditional in the ResponsiveLerpToTargetBehaviour that was
preventing responsive actions from firing when deployed to devices
-SortingOrder and SortingLayer buttons for EditTMProComponent
-Other ResponsiveElement fixes