Add functionality to configure sequence autoplaying and axis switches via Timeline
Added the new SequenceAutoplayConfig and SequencePauseMomentumconfig that,
when given a Sequence, will set the autoplay and pause momentum thresholds
directly from Timeline. This should save a lot of assembly time, as we
no longer need to do this manually for each slot, going back and forth
between Timeline and scriptable object. It also means that the autoplay
thresholds can be manipulated with the other timeline utilities used for
multiplying clips, setting time, etc.

Also did a similar thing with Axis Switches. Note that the InvertSwitch
hasn't been polished up to work with the new functionality, so there might
be some snags there.