Add TextUtils editor tool; begin adding functionality for hotkeys
Began adding quality-of-life tools to speed up assembly time even
more.

The TextUtils tool is one that allows the user to use a text object
as an onion skin to figure out placement for other text objects. It can
be enabled and disabled via hotkeys, as well as populated with the content
of selected text objects with the click of a button. This is really useful
for figuring out, for example, rect transform placement, and having those
values stored with the onion skin so they aren't lost while placing a
particular sequence.

Importantly, TextUtils also includes methods for extracting text from
objects in Unity, copying them to the clipboard. This is useful for
when you need to modify the line lengths of a storyboard and need to
generate data for an XML file in the correct format quickly. Simply
input the text in question into the, say, the onion skin, set the width
and other parameters accordingly, then extract the text with the
appropriate wrapper.

Other than that, began abstracting logic within some editor tools to
allow them to be hooked up to the Unity Shortcuts functionality. This
doesn't require much work, since most or all of the methods in the
editor tools are already static (they just need the appropriate MenuItem
attribute to get hooked in). They may also require a reference to the
variables inside the editor window in question; to do this, you just need
to create a static reference to the window so a static method can access any
dependent instance variables in its parent class.

Will continue adding hotkey functionality as needed. Next steps are to
continue adding quality of life tools to further improve assembly workflow.