Revise ScrollSnapController and add related tools; fix sequence scrubbing
Two big changes. Cleaned up the ScrollSnapController and made it so
it works outside of the UI Canvas system and instead works with the
regular 2D collider system. Also added related editor tools to allow
us to easily create new panels, as well as respace the icons associated
with each panel.

The other change was to fix the sequence scrubbing and actually make it
functional. To do this, I needed to revise the playables by adding some
cleanup code OnGraphStop in the mixer behaviours. While the application
is playing, we'll want timelines that get deactivated to set their content
to transparent so it doesn't get in the way of active timeline animations.
The reason we need to do this is because, when scrubbing, since the user
moves through the timeline so quickly, many frames are skipped, which
means that the transitions between timelines often come out incorrectly.

That being said, we only need to execute this timeline cleanup code when
scrubbing, so I also added other helper attributes to the mixer behaviours,
including including a reference to the scriptable object bool IsScrubbing.
We only want to execute this code when scrubbing because, when progressing
through the timeline slowly, via swiping, the reset code will cause the screen
to flash to white for brief instant.

Next steps are to clean up the scrubbing code, then move onto continuing work
on editor tools.