WORK IN PROGRESS - Begin overhauling sequence functionality
Made a lot of big changes to the way sequence data is manipulated.
At a high level, a MasterSequence Monobehaviour, taking the place of
SequenceList, now holds reference to a list of Sequences, as well as a
number of InputControllers. These InputControllers include types such
as AutorunController, NavigationController, and TouchController, which
are used to house and manage large interdependent swathes of functionality
related to manipulating sequences.

For example, the AutorunController contains references to the Autoplay and
Lerper input modules, which both make use of predefined autoplay intervals,
autoplay configuration markers, related utility functions used to determine
if an interval is active or can be updated, etc. The AutorunController
then, creates its own representation of the data inside of MasterSequence
so that its input modules can function correctly internally, while also being
able to interface with any logic at the master level in a modular way. The idea
is that each input controller should have its own namespace and be able to
operate independently of the other input controllers. All of the Autorun functionality,
for example, is housed inside of the AltSalt.Sequencing.Autorun namespace,
whereas all of the Navigation functionality is housed inside of the
AltSalt.Sequencing.Navigation namespace.

I still need to move some dependencies over into the proper namespaces, but wanted
to do this commit before overhauling the folder structure further.

Regarding the TouchController, it works similarly to the AutorunController, just
that it's a bit more complex. The TouchController contains the input modules
AxisModifier, SwipeModifier, and MomentumModifier, which together handle all
input that could modify a sequence using a swipe mechanic. Importantly, the AxisModifier
contains a variety of axis switches; each handles the activation or deactivation
of a particular axis (ySwipeAxis, xMomentumAxis) in its own way, which of course can have
consequences for how all input is resolved on a sequence, as well as, in the case of the
ForkSwitch, consequences for what happens when the end of a sequence is reached.

In order to accommodate this, not only does the TouchController create its
own representation of the MasterSequence data for use across input modules, its child
input module, Axis Modifie, contains axis switches which also creates
their own internal configuration data that they use to propagate config
changes upstream to the TouchController and MasterSequence.

To emphasize that these individual parts are all sections of an interdependent system,
I've broken the AxisModifier class into partial classes (will continue thinking
through whether there's a better way to handle this).

Other than that, it's worth noting that the AutorunController, TouchController, and
AxisModifier, holding reference to sequences via MasterSequence, all reach into
sequences' associated PlayableAssets to pull configuration data
from autoplay markers, pause momentum markers, and axis confiugration clips.

In addiition, it's worth noting that the AxisModifier in particular
makes use of the MasterSequence's knowledge of global time, spanning all
sequences, to store the activation times of axis switches in global time as well. The idea
is that, even when jumping abruptly from one part of a story to another, or scrubbing through
a story quickly, we will always be able to compare a axis switch's activation time to the
global time and set all touch input variables (ySwipeAxis, xMomentumAxis, etc.) correctly.

Lastly, the idea is that, coded correctly, we should be able to continue creating additional
kinds of InputControllers to manipulate sequences without breaking existing functionality.
For example, it would be possible to create a TimerController to advance a set of sequences
based on an external clock, or a LightController to advance sequences based on light input,
a GyroscopeController to manipulate sequences using device tilt, perhaps a SoundController
to manipulate sequences using different kinds of sound input, etc. With any luck,
this can also help protect against changes in technology for the future, allowing
us to extend this core idea as new kinds of input and paradigms emerge.