WORK IN PROGRESS - Finish first pass at sequence refactor; restructure part of the project
Fixed all of the compile errors related to the sequence refactor. Also restructured folder
setup so that it mirrors standard namespacing schemas.

Moved all of the logic for activating and deactivating adjacent sequences to the
new ProcessModifySequence script. Sequences also now contain references to the SequenceConfig
Monobehaviour, which in turn contains references to all of the components / functionality
you would need in order to work with a sequence, like ProcessModifySequence or
SyncTimelineToSequence.

Still need to clean up a lot of commented code and clean up the implementation for
handling adjacent sequences, though. Important to note is that this refactor still
hasn't been tested in the editor; I'm just committing now so that any changes
from here on out can be recorded and reverted if need be (the rearrangment of the
folders prevents you from seeing changes correctly unless the files have been
committed).

Next step is to run this in the editor, reorganize scenes, dealing with any major
issues and polishing up the implementation.