WORK IN PROGRESS - Continue fixing sequence refactor details; begin refactoring axis switches
Managed to fix some issues with the sequence refactor, including
the input modules not being registered correctly to the Master Sequence.
The modules all now register themselves to the Master Sequence by passing their
names during the requestApplyModify event, though this should probably be
changed to use a more unique identifier.

Other than that, fixed a problem with MomentumModifier in that it was interfering
with the axis switches when modifying the IsReversing variable.

Importantly, I realized that the axis switches are buggy and rather than
try to fix the current implementation, I'm currently investigating how to make
the axis switches and branching paths robust so that any input module can make
use of them. Lots of half-baked ideas being committed at the moment, but
I'm about to move some files around and break up some of the functionality
inside of the Sequencing/Touch folder, so want to get a snapshot of the
functionality before I do.

Next step is to move all of the branching logic to the root of the Sequencing
folder, refactor the Touch functionality to extend that, and move toward
finalizing the axis switches.